Move build process to composer scripts
- Allows simplifying AWS build script
- Allows easily running build locally

+++ src/Console/ClearCache.php
@@ -33,6 +33,10 @@ class ClearCache
                 continue;
             }
 
+            if ($file->getFilename() === '.placeholder') {
+                continue;
+            }
+
             unlink($file->getRealPath());
         }
